+++ adapters/swift/AppStylrTokens.swift
@@ -3,7 +3,7 @@
 import SwiftUI
 
 public enum AppStylrTokens {
-  public static let version = "1.0.0"
+  public static let version = "1.1.0"
 
   public enum Brand {
     public static let ink = Color(red: 0.066667, green: 0.066667, blue: 0.066667)